ICU4N.Globalization: Added ThreadCultureChange for testing purposes
